+++ src/abiEncoderV2/calldata_array_short_no_revert_string.sol
@@ -1,8 +1,3 @@
 contract C {
-	function f(uint[] calldata) public {}
+    function f(uint256[] calldata) public {}
 }
-// ----
-// f(uint256[]): 0x20, 0 ->
-// f(uint256[]): 0x20, 1 -> FAILURE
-// f(uint256[]): 0x20, 2 -> FAILURE
-
