+++ src/dataLocations/internalFunction/internal_function_parameters_no_data_location.sol
@@ -1,5 +1,3 @@
 contract C {
-    function g(uint[]) internal pure {}
+    function g(uint256[]) internal pure {}
 }
-// ----
-// TypeError 6651: (28-34): Data location must be "storage", "memory" or "calldata" for parameter in function, but none was given.
