+++ src/dataLocations/internalFunction/internal_function_return_parameters_no_data_location.sol
@@ -1,5 +1,3 @@
 contract C {
-    function g() internal pure returns(uint[]) {}
+    function g() internal pure returns (uint256[]) {}
 }
-// ----
-// TypeError 6651: (52-58): Data location must be "storage", "memory" or "calldata" for return parameter in function, but none was given.
