+++ src/immutable/non-value_type.sol
@@ -1,5 +1,3 @@
 contract C {
-    uint[] immutable x;
+    uint256[] immutable x;
 }
-// ----
-// TypeError 6377: (17-35): Immutable variables cannot have a non-value type.
