+++ src/modifiers/evaluation_order.sol
@@ -1,20 +1,26 @@
-contract A { constructor(uint) {} }
-contract B { constructor(uint) {} }
-contract C { constructor(uint) {} }
+contract A {
+    constructor(uint256) {}
+}
+
+contract B {
+    constructor(uint256) {}
+}
+
+contract C {
+    constructor(uint256) {}
+}
 
 contract D is A, B, C {
-    uint[] x;
+    uint256[] x;
     constructor() m2(f(1)) B(f(2)) m1(f(3)) C(f(4)) m3(f(5)) A(f(6)) {
         f(7);
     }
 
-    function query() public view returns (uint[] memory) { return x; }
+    function query() public view returns (uint256[] memory) {return x;}
 
-    modifier m1(uint) { _; }
-    modifier m2(uint) { _; }
-    modifier m3(uint) { _; }
+    modifier m1(uint256) {_;}
+    modifier m2(uint256) {_;}
+    modifier m3(uint256) {_;}
 
-    function f(uint y) internal returns (uint) { x.push(y); return 0; }
+    function f(uint256 y) internal returns (uint256) {x.push(y);return 0;}
 }
-// ----
-// query() -> 0x20, 7, 4, 2, 6, 1, 3, 5, 7
