+++ src/modifiers/modifier_overrides_variable.sol
@@ -1,5 +1,7 @@
-contract A { modifier mod(uint a) { _; } }
-contract B is A { uint public mod; }
-// ----
-// DeclarationError 9097: (61-76): Identifier already declared.
-// TypeError 1456: (61-76): Override changes modifier to public state variable.
+contract A {
+    modifier mod(uint256 a) {_;}
+}
+
+contract B is A {
+    uint256 public mod;
+}
