+++ src/nameAndTypeResolution/559_no_warning_for_using_members_that_look_like_address_members.sol
@@ -1,7 +1,7 @@
 abstract contract C {
-    function transfer(uint) public virtual;
+    function transfer(uint256) public virtual;
+
     function f() public {
         this.transfer(10);
     }
 }
-// ----
