+++ src/natspec/docstring_author_title_state_variable.sol
@@ -1,8 +1,5 @@
 contract C {
     /// @title title
     /// @author author
-    uint private state;
+    uint256 private state;
 }
-// ----
-// DocstringParsingError 6546: (17-56): Documentation tag @author not valid for non-public state variables.
-// DocstringParsingError 6546: (17-56): Documentation tag @title not valid for non-public state variables.
