+++ src/natspec/docstring_named_return_parameter.sol
@@ -1,5 +1,4 @@
 abstract contract C {
     /// @return value The value returned by this function.
-    function vote() public virtual returns (uint value);
+    function vote() public virtual returns (uint256 value);
 }
-// ----
