+++ src/natspec/docstring_non_public_state_variable_with_return.sol
@@ -1,6 +1,4 @@
 contract test {
-  /// @return returns something
-  uint private state;
+    /// @return returns something
+    uint256 private state;
 }
-// ----
-// DocstringParsingError 6546: (18-47): Documentation tag @return not valid for non-public state variables.
