+++ src/natspec/docstring_state_variable.sol
@@ -1,6 +1,5 @@
 contract C {
     /// @notice example of notice
     /// @dev example of dev
-    uint public state;
+    uint256 public state;
 }
-// ----
