+++ src/types/mapping/contract_storage_parameter_with_mapping.sol
@@ -1,6 +1,7 @@
-struct S { mapping(uint => uint)[2] a; }
+struct S {
+    mapping(uint256 => uint256)[2] a;
+}
+
 contract C {
     function f(S storage s) public {}
 }
-// ----
-// TypeError 6651: (69-80): Data location must be "memory" or "calldata" for parameter in function, but "storage" was given.
