+++ src/types/struct_mapping_abstract_constructor_param.sol
@@ -1,24 +1,19 @@
 struct S {
-	mapping (uint => uint) m;
+    mapping(uint256 => uint256) m;
 }
 
 abstract contract A {
-	constructor (S storage s) {
-		s.m[5] = 16;
-	}
+    constructor(S storage s) {
+        s.m[5] = 16;
+    }
 }
 
 contract C is A {
-	mapping(uint => S) m;
+    mapping(uint256 => S) m;
 
-	constructor() A(m[1]) {
-	}
+    constructor() A(m[1]) {}
 
-	function getM(uint a, uint b) external returns (uint) {
-		return m[a].m[b];
-	}
+    function getM(uint256 a, uint256 b) external returns (uint256) {
+        return m[a].m[b];
+    }
 }
-// ----
-// getM(uint256,uint256): 0, 0 -> 0
-// getM(uint256,uint256): 1, 5 -> 0x10
-// getM(uint256,uint256): 1, 0 -> 0
