+++ src/types/type_conversion_cleanup.sol
@@ -1,7 +1,6 @@
 contract Test {
-    function test() public returns (uint ret) { return uint(uint160(address(uint160(uint128(type(uint200).max))))); }
+    function test()
+        public
+        returns (uint256 ret)
+    {return uint(uint160(address(uint160(uint128(type(uint200).max)))));}
 }
-// ====
-// compileToEwasm: also
-// ----
-// test() -> 0xffffffffffffffffffffffffffffffff
